Add options to store event weights
Event weights require significant storage and the map makes it hard to use with python. Options on how to run and store weights have been added to allow for different use cases.
Event weights can now be stored in the common tree, in a dedicated tree, or both. In the common tree, all the weights within a given category (flux, Genie, hadron reinteractions) can be combined into one and stored as a vector instead of a map. The combined Genie and the combined reinteraction weights use simultaneous multisim over all individual weights while the combined flux weight is the product of the individual multisim weights. Options can be toggled by commenting or uncommenting lines in the MC fcl file.
By default, Genie weights are not run, no dedicated tree is created, the map is not stored in the common tree and vectors for combined flux and combined reinteractions weights are created.